feat(CRUD): Prospects CRUDs
Made the createProspect() CRUD and the analyseAndSanitiseProspect() functions.

analyseAndSanitizeProspect() : This function will be usable by many other function, his role is to check if the 'prospect' object correspond to all of the exigences like :
- All the fields are present and not null
- All the fields have the good type
- Sector Id is existant
- Email and phone good format
His other role is to 'Title case' the names and city fields
If one of the analyses fail, it will return an Error Object

createProspect() : This CRUD function will take all the content of the Prospect DB file to get the JSON object of it and add the prospect into it.
After that we write the file.

I also maked the logs more readable in prospectsCRUD.js and sectorsCRUD.js